Move shared utilities into lib/, paths to top-level, ocean to lib/
Phase 1 of codebase restructure. Move scattered utility modules from
app/ into a centralized lib/ package: alerts, auth, context, deps,
events, exceptions, guards, ocean, permissions, request_type,
response_type. Move paths.py to package top-level. Consolidate
dependency factories (event_with, game_with, party_with, time_slot_with)
into lib/deps.py. Fix pre-existing runtime bug in request_type.py where
State and Litestar were not imported in the else branch.
