+++ convergence_games/app/app_config/exception_handlers.py
@@ -1,11 +1,11 @@
 from litestar.datastructures import Cookie
 from litestar.response import Redirect
 
-from convergence_games.app.alerts import AlertError
 from convergence_games.app.app_config.template_config import catalog
-from convergence_games.app.exceptions import UserNotLoggedInError
-from convergence_games.app.request_type import Request
-from convergence_games.app.response_type import HTMXBlockTemplate
+from convergence_games.lib.alerts import AlertError
+from convergence_games.lib.exceptions import UserNotLoggedInError
+from convergence_games.lib.request_type import Request
+from convergence_games.lib.response_type import HTMXBlockTemplate
 
 
 def user_not_logged_in_handler(request: Request, exc: UserNotLoggedInError) -> Redirect:
