+++ convergence_games/app/routers/frontend/editor_test.py
@@ -1,7 +1,7 @@
 from litestar import Controller, get
 
-from convergence_games.app.request_type import Request
-from convergence_games.app.response_type import HTMXBlockTemplate, Template
+from convergence_games.lib.request_type import Request
+from convergence_games.lib.response_type import HTMXBlockTemplate, Template
 
 
 class EditorTestController(Controller):
