+++ convergence_games/app/routers/frontend/event_manager.py
@@ -23,13 +23,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import aliased, selectinload, with_loader_criteria
 from sqlalchemy.sql.selectable import Select
 
-from convergence_games.app.alerts import Alert, AlertError
 from convergence_games.app.app_config.template_config import catalog
-from convergence_games.app.context import user_id_ctx
-from convergence_games.app.guards import permission_check, user_guard
-from convergence_games.app.request_type import Request
-from convergence_games.app.response_type import HTMXBlockTemplate
-from convergence_games.app.routers.frontend.common import event_with
 from convergence_games.db.enums import GameClassification, SubmissionStatus, TimeSlotStatus
 from convergence_games.db.models import (
     Allocation,
@@ -50,8 +44,14 @@ from convergence_games.db.models import (
     UserGamePlayed,
     UserGamePreference,
 )
-from convergence_games.db.ocean import Sqid, sink, swim
-from convergence_games.permissions import user_has_permission
+from convergence_games.lib.alerts import Alert, AlertError
+from convergence_games.lib.context import user_id_ctx
+from convergence_games.lib.deps import event_with
+from convergence_games.lib.guards import permission_check, user_guard
+from convergence_games.lib.ocean import Sqid, sink, swim
+from convergence_games.lib.permissions import user_has_permission
+from convergence_games.lib.request_type import Request
+from convergence_games.lib.response_type import HTMXBlockTemplate
 from convergence_games.services.algorithm.game_allocator import (
     AlgPartyP,
     GameAllocator,
