+++ convergence_games/app/routers/frontend/misc_components.py
@@ -1,7 +1,7 @@
 from litestar import Controller, get
 
 from convergence_games.app.app_config.template_config import catalog
-from convergence_games.app.response_type import HTMXBlockTemplate, Template
+from convergence_games.lib.response_type import HTMXBlockTemplate, Template
 
 
 class MiscComponentsController(Controller):
