+++ convergence_games/app/routers/frontend/my_submissions.py
@@ -5,10 +5,10 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
-from convergence_games.app.guards import user_guard
-from convergence_games.app.request_type import Request
-from convergence_games.app.response_type import HTMXBlockTemplate
 from convergence_games.db.models import Event, Game
+from convergence_games.lib.guards import user_guard
+from convergence_games.lib.request_type import Request
+from convergence_games.lib.response_type import HTMXBlockTemplate
 from convergence_games.settings import SETTINGS
