+++ convergence_games/app/routers/frontend/profile.py
@@ -11,19 +11,19 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from convergence_games.app.app_config.jwt_cookie_auth import build_token_extras, jwt_cookie_auth
-from convergence_games.app.common.auth import (
+from convergence_games.db.enums import LoginProvider
+from convergence_games.db.models import User, UserEventRole, UserLogin
+from convergence_games.lib.auth import (
     AuthIntent,
     OAuthRedirectState,
     ProfileInfo,
     authorize_flow,
 )
-from convergence_games.app.events import EVENT_EMAIL_SIGN_IN
-from convergence_games.app.guards import user_guard
-from convergence_games.app.request_type import Request
-from convergence_games.app.response_type import HTMXBlockTemplate, Template
-from convergence_games.db.enums import LoginProvider
-from convergence_games.db.models import User, UserEventRole, UserLogin
-from convergence_games.db.ocean import Sqid, sink
+from convergence_games.lib.events import EVENT_EMAIL_SIGN_IN
+from convergence_games.lib.guards import user_guard
+from convergence_games.lib.ocean import Sqid, sink
+from convergence_games.lib.request_type import Request
+from convergence_games.lib.response_type import HTMXBlockTemplate, Template
 from convergence_games.utils.email import normalize_email
