+++ convergence_games/app/routers/static/static.py
@@ -1,7 +1,7 @@
 from litestar.datastructures import CacheControlHeader
 from litestar.static_files import create_static_files_router
 
-from convergence_games.app.paths import STATIC_DIR_PATH
+from convergence_games.paths import STATIC_DIR_PATH
 
 static_files_router = create_static_files_router(
     path="/static",
