+++ convergence_games/lib/exceptions.py
