+++ convergence_games/lib/permissions.py
