+++ convergence_games/permissions/__init__.py
@@ -1,4 +1,4 @@
-from .permissions import user_has_permission
+from convergence_games.lib.permissions import user_has_permission
 
 __all__ = [
     "user_has_permission",
