+++ tests/app/common/test_auth.py
@@ -8,7 +8,7 @@ import pytest
 import pytest_asyncio
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
-from convergence_games.app.common.auth import (
+from convergence_games.lib.auth import (
     AccountAlreadyExistsError,
     AuthIntent,
     NoAccountForSignInError,
